city/requirements.txt
asgiref==3.8.1
Django==5.0.7
djangorestframework==3.15.2
psycopg2-binary==2.9.9
sqlparse==0.5.1
tzdata==2024.1
